+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 - **Tables / Media Previews**: Added tooltips with filename, size and modification date to all previews.
 - **Tables / Media Recorder**: Added new "expert" mode which allows to customize the ffmpeg.exe command for every screen. Please refer to the ffmpeg documentation for more details. And if someone manages to record with audio, please share the corresponding command on my Discord!
 - **Competitions / iScored**: Added error message in case the game room's read API is not enabled.
+- **DMD Position Tool**: Added ability to turn external DMD off in VpinMame settings and/or disable DMD in DmdDevice.ini.
 
 
 ## Bugfixes
@@ -15,6 +16,7 @@
 - **Tables / Media Recorder**: Added "move back" behaviour when the recording is started from the client on the cabinet. This should solve the issue that the Studio stays in front during the recording.
 - **Webhooks**: Added missing change listener for the game lifecycle events (create, update and delete).
 - **Backglass Manager / .res Editor**: Fixed saving .res files without frame images.
+- **Backglass Manager / Screen offset**: Fixed the backglass screen identification when parsing screenres.txt.
 - **Highscore Event Handling**: Fixed issue that if the player achieved multiple new highscores during one game, only one highscore change event was emitted. The server emits now messages for every change, starting with the highest one. This way, all new scores are posted to iScored too (or only the highest if the single score submission is enabled).
 - **Future Pinball Support**: Phew, this was neglected a lot. So here comes a bunch of fixes based on table archives from **Terry Red**. I still lack some knowledge here, so feel free to drop some feedback on the VPin Studio's Discord server.
   - Fixed auto-renaming for .fpt files.
@@ -22,6 +24,7 @@
   - Added renaming support for BAM configuration files.
   - Added upload detection for BAM configuration files, including a separate upload dialog.
   - Added option in the table deletion dialog for BAM configuration files.
+
 
 ## VPin Mania